dark-CleanupSync

+++ src/dark/core/client/renders/RenderBlockFluid.java
@@ -54,9 +54,9 @@ public class RenderBlockFluid
         }
 
         int color = fluidstack.getFluid().getColor(fluidstack);
-        float red = (float) (color >> 16 & 255) / 255.0F;
-        float green = (float) (color >> 8 & 255) / 255.0F;
-        float blue = (float) (color & 255) / 255.0F;
+        float red = (color >> 16 & 255) / 255.0F;
+        float green = (color >> 8 & 255) / 255.0F;
+        float blue = (color & 255) / 255.0F;
         GL11.glColor4f(red, green, blue, 1);
     }
