+++ src/dark/core/prefab/invgui/InvChest.java
@@ -182,7 +182,7 @@ public class InvChest implements IInvBox
     @Override
     public boolean isUseableByPlayer(EntityPlayer par1EntityPlayer)
     {
-        return this.hostTile.worldObj.getBlockTileEntity(this.hostTile.xCoord, this.hostTile.yCoord, this.hostTile.zCoord) != this.hostTile ? false : par1EntityPlayer.getDistanceSq((double) this.hostTile.xCoord + 0.5D, (double) this.hostTile.yCoord + 0.5D, (double) this.hostTile.zCoord + 0.5D) <= 64.0D;
+        return this.hostTile.worldObj.getBlockTileEntity(this.hostTile.xCoord, this.hostTile.yCoord, this.hostTile.zCoord) != this.hostTile ? false : par1EntityPlayer.getDistanceSq(this.hostTile.xCoord + 0.5D, this.hostTile.yCoord + 0.5D, this.hostTile.zCoord + 0.5D) <= 64.0D;
     }
 
     @Override
